storage/requirements.txt
pydantic>=2.0
python-dotenv
psycopg2-binary>=2.9.9
psycopg2-pool>=1.1
